update code structure and support m2 zero build

+++ scripts/bootloader.sh
@@ -1,23 +1,21 @@
 #!/bin/bash
 
-#gunzip -c BPI_M3_1080P.img.gz | dd of=/dev/mmcblk0 conv=sync,noerror bs=1k
-
 die() {
         echo "$*" >&2
         exit 1
 }
 
 [ -s "./env.sh" ] || die "please run ./configure first."
-. ./env.sh
-O=$1
 
+. ./env.sh
+
+O=$1
 if [ ! -z $O ] ; then
 	BOARD=$O
 fi
 
-P=${TOPDIR}/output/$BOARD/pack
+P=${TOPDIR}/output/${BOARD}/pack
 U=${TOPDIR}/output/100MB
-#echo DOWNLOAD DIR=$P
 
 mkdir -p $U
 TMP_FILE=${U}/${BOARD}.tmp
@@ -34,12 +32,14 @@ LOOP_DEV=`sudo losetup -f --show ${TMP_FILE}`
 #(sudo dd if=$P/boot.fex 	of=${LOOP_DEV} bs=1k seek=86016) >/dev/null 2>&1
 
 sudo sync
-
 sudo losetup -d ${LOOP_DEV}
 
 (dd if=${TMP_FILE} of=${IMG_FILE} bs=1024 skip=8 count=102392 status=noxfer) >/dev/null 2>&1
 
-rm -f ${IMG_FILE}.gz
+if [ -f ${IMG_FILE}.gz ]; then
+	rm -f ${IMG_FILE}.gz
+fi
+
 echo "gzip ${IMG_FILE}"
 gzip ${IMG_FILE}
-rm -f ${TMP_FILE}
+sudo rm -f ${TMP_FILE}
